correctly populate user object by default
This ensures that everywhere that uses .req.user wil also have access to a
fully functional Model fo req.user.roles.account.sites[]....
This is crucial for ensuring any testing with the global sesssion/cookie will
work, as it relies on the record from req.user.roles.account.sites[n].key.
Without this, the record is just an internal reference to an ObjectId in the
site key itself.