onboarding wizard && better dexcom share integration
Make the Dexcom Share experience smoother by providing a clear way to delete
all Dexcom credentials.

Provide stub experience for "new site" creation which includes tracking
tour_given_at flag on the Site document, and invoking a modal if it is absent.
Engaging with the UI or dismissing with "don't bother me again" checked
properly sets the flag.

This causes the modal to appear immediately after creating a site during the
onboarding experience, and not otherwise.